+++ heylogs-api/src/main/java/internal/heylogs/SemverRule.java
@@ -4,44 +4,52 @@ import com.vladsch.flexmark.ast.Heading;
 import com.vladsch.flexmark.util.ast.Node;
 import nbbrd.design.VisibleForTesting;
 import nbbrd.heylogs.Failure;
-import nbbrd.heylogs.Rule;
 import nbbrd.heylogs.Version;
+import nbbrd.heylogs.spi.Rule;
 import nbbrd.service.ServiceProvider;
+import org.jetbrains.annotations.NotNull;
 import org.semver4j.Semver;
 
 @ServiceProvider
 public final class SemverRule implements Rule {
 
     @Override
-    public String getName() {
+    public @NotNull String getId() {
         return "semver";
     }
 
     @Override
-    public Failure validate(Node node) {
-        return node instanceof Heading ? validateSemVer((Heading) node) : null;
+    public Failure validate(@NotNull Node node) {
+        return node instanceof Heading ? validateSemVer((Heading) node) : NO_PROBLEM;
     }
 
     @Override
     public boolean isAvailable() {
-        return Rule.isEnabled(System.getProperties(), getName());
+        return Rule.isEnabled(System.getProperties(), getId());
     }
 
     @VisibleForTesting
     Failure validateSemVer(Heading heading) {
         if (!Version.isVersionLevel(heading)) {
-            return null;
+            return NO_PROBLEM;
         }
 
         try {
             Version version = Version.parse(heading);
             if (version.isUnreleased()) {
-                return null;
+                return NO_PROBLEM;
             }
             String ref = version.getRef();
-            return Semver.isValid(ref) ? null : Failure.of(this, "Invalid semver format: '" + ref + "'", heading);
+            return Semver.isValid(ref)
+                    ? NO_PROBLEM
+                    : Failure
+                    .builder()
+                    .rule(this)
+                    .message("Invalid semver format: '" + ref + "'")
+                    .location(heading)
+                    .build();
         } catch (IllegalArgumentException ex) {
-            return null;
+            return NO_PROBLEM;
         }
     }
 }