+++ heylogs-cli/src/main/java/nbbrd/heylogs/cli/CheckCommand.java
@@ -1,19 +1,17 @@
 package nbbrd.heylogs.cli;
 
-import internal.heylogs.cli.FailureFormatOptions;
+import internal.heylogs.SemverRule;
+import internal.heylogs.StylishFormat;
+import internal.heylogs.cli.FormatCandidates;
 import internal.heylogs.cli.MarkdownInputSupport;
-import internal.heylogs.cli.RuleSetOptions;
 import nbbrd.console.picocli.FileOutputOptions;
 import nbbrd.console.picocli.MultiFileInputOptions;
-import nbbrd.heylogs.Failure;
-import nbbrd.heylogs.FailureFormatter;
-import nbbrd.heylogs.Rule;
+import nbbrd.heylogs.Checker;
 import picocli.CommandLine;
 import picocli.CommandLine.Command;
 
 import java.io.Writer;
 import java.nio.file.Path;
-import java.util.List;
 import java.util.concurrent.Callable;
 
 import static internal.heylogs.cli.MarkdownInputSupport.newMarkdownInputSupport;
@@ -28,29 +26,48 @@ public final class CheckCommand implements Callable<Void> {
     @CommandLine.Mixin
     private FileOutputOptions output;
 
-    @CommandLine.Mixin
-    private RuleSetOptions ruleSet;
+    @CommandLine.Option(
+            names = {"-s", "--semver"},
+            defaultValue = "false",
+            description = "Mention if this changelog follows Semantic Versioning."
+    )
+    private boolean semver;
 
-    @CommandLine.Mixin
-    private FailureFormatOptions format;
+    @CommandLine.Option(
+            names = {"-f", "--format"},
+            paramLabel = "<name>",
+            defaultValue = StylishFormat.ID,
+            description = "Specify the format used to control the appearance of the result. Valid values: ${COMPLETION-CANDIDATES}.",
+            completionCandidates = FormatCandidates.class
+    )
+    private String formatId;
 
     @Override
     public Void call() throws Exception {
         try (Writer writer = newTextOutputSupport().newBufferedWriter(output.getFile())) {
 
-            List<Rule> rules = ruleSet.getRules();
-            FailureFormatter formatter = format.getFormatter();
+            Checker checker = getChecker();
             MarkdownInputSupport markdown = newMarkdownInputSupport();
 
             for (Path file : input.getAllFiles(markdown::accept)) {
-                formatter.format(
+                checker.formatFailures(
                         writer,
                         markdown.getName(file),
-                        Failure.allOf(markdown.readDocument(file), rules)
+                        checker.validate(markdown.readDocument(file))
                 );
             }
         }
 
         return null;
+    }
+
+    private Checker getChecker() {
+        Checker.Builder result = Checker.ofServiceLoader()
+                .toBuilder()
+                .formatId(formatId);
+        if (semver) {
+            result.rule(new SemverRule());
+        }
+        return result.build();
     }
 }