+++ heylogs-maven-plugin/src/main/java/nbbrd/heylogs/maven/plugin/CheckMojo.java
@@ -1,34 +1,35 @@
 package nbbrd.heylogs.maven.plugin;
 
-import com.vladsch.flexmark.parser.Parser;
 import com.vladsch.flexmark.util.ast.Document;
+import internal.heylogs.SemverRule;
+import internal.heylogs.StylishFormat;
+import nbbrd.heylogs.Checker;
 import nbbrd.heylogs.Failure;
-import nbbrd.heylogs.Rule;
-import nbbrd.heylogs.RuleLoader;
-import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
 import org.semver4j.Semver;
 
+import java.io.BufferedReader;
 import java.io.File;
 import java.io.IOException;
-import java.io.Reader;
-import java.nio.file.Files;
+import java.io.StringReader;
 import java.util.List;
 
+import static java.util.Locale.ROOT;
+
 @Mojo(name = "check", defaultPhase = LifecyclePhase.VALIDATE, threadSafe = true)
-public final class CheckMojo extends AbstractMojo {
+public final class CheckMojo extends HeylogsMojo {
 
     @Parameter(defaultValue = "${project.basedir}/CHANGELOG.md", property = "heylogs.input.file")
     private File inputFile;
 
-    @Parameter(defaultValue = "false", property = "heylogs.skip")
-    private boolean skip;
-
     @Parameter(defaultValue = "false", property = "heylogs.semver")
     private boolean semver;
+
+    @Parameter(defaultValue = StylishFormat.ID, property = "heylogs.format.id")
+    private String formatId;
 
     @Parameter(defaultValue = "${project.version}", readonly = true)
     private String projectVersion;
@@ -44,70 +45,56 @@ public final class CheckMojo extends AbstractMojo {
         }
 
         if (semver) {
-            enableSemanticVersioning();
+            checkSemanticVersioning();
         }
 
         if (inputFile.exists()) {
-            validateFile();
+            check(loadChecker());
         } else {
-            if (isRootProject()) {
-                raiseErrorMissingFile();
+            if (isRootProject(projectBaseDir)) {
+                raiseErrorMissingChangelog();
             } else {
-                notifyMissingFile();
+                notifyMissingChangelog();
             }
         }
     }
 
-    private void enableSemanticVersioning() throws MojoExecutionException {
+    private void checkSemanticVersioning() throws MojoExecutionException {
         getLog().info("Using Semantic Versioning specification");
         if (Semver.isValid(projectVersion)) {
             getLog().info("Valid project version");
-            System.setProperty(Rule.ENABLE_KEY, "semver");
         } else {
-            getLog().error(String.format("Invalid project version: '%s' must follow Semantic Versioning specification (https://semver.org/)", projectVersion));
+            getLog().error(String.format(ROOT, "Invalid project version: '%s' must follow Semantic Versioning specification (https://semver.org/)", projectVersion));
             throw new MojoExecutionException("Invalid project version. See above for details.");
         }
     }
 
-    private void validateFile() throws MojoExecutionException {
+    private Checker loadChecker() {
+        Checker.Builder result = Checker.ofServiceLoader()
+                .toBuilder()
+                .formatId(formatId);
+        if (semver) {
+            result.rule(new SemverRule());
+        }
+        return result.build();
+    }
+
+    private void check(Checker checker) throws MojoExecutionException {
+        Document changelog = readChangelog(inputFile);
+        List<Failure> failures = checker.validate(changelog);
+        writeFailures(checker, failures);
+        if (!failures.isEmpty()) {
+            throw new MojoExecutionException("Invalid changelog");
+        }
+    }
+
+    private void writeFailures(Checker checker, List<Failure> failures) throws MojoExecutionException {
         try {
-            getLog().info("Reading " + inputFile);
-            Document changelog = read();
-
-            List<Failure> failures = Failure.allOf(changelog, RuleLoader.load());
-            if (!failures.isEmpty()) {
-                getLog().error("Invalid changelog");
-                failures.forEach(failure -> getLog().error(failure.toString()));
-                throw new MojoExecutionException("Invalid changelog");
-            }
-            getLog().info("Valid changelog");
+            StringBuilder text = new StringBuilder();
+            checker.formatFailures(text, inputFile.toString(), failures);
+            new BufferedReader(new StringReader(text.toString())).lines().forEach(!failures.isEmpty() ? getLog()::error : getLog()::info);
         } catch (IOException ex) {
-            throw new MojoExecutionException("Error while checking changelog", ex);
-        }
-    }
-
-    private boolean isRootProject() {
-        File parentDir = projectBaseDir.getParentFile();
-        if (parentDir != null) {
-            File parentPom = new File(parentDir, "pom.xml");
-            return !parentPom.exists();
-        }
-        return true;
-    }
-
-    private void raiseErrorMissingFile() throws MojoExecutionException {
-        getLog().error("Missing changelog");
-        throw new MojoExecutionException("Missing changelog");
-    }
-
-    private void notifyMissingFile() {
-        getLog().info("Changelog not found");
-    }
-
-    public Document read() throws IOException {
-        Parser parser = Parser.builder().build();
-        try (Reader reader = Files.newBufferedReader(inputFile.toPath())) {
-            return parser.parseReader(reader);
+            throw new MojoExecutionException("Error while writing failures", ex);
         }
     }
 }