+++ heylogs-maven-plugin/src/main/java/nbbrd/heylogs/maven/plugin/ExtractMojo.java
@@ -1,29 +1,21 @@
 package nbbrd.heylogs.maven.plugin;
 
-import com.vladsch.flexmark.formatter.Formatter;
-import com.vladsch.flexmark.parser.Parser;
 import com.vladsch.flexmark.util.ast.Document;
+import nbbrd.heylogs.Extractor;
 import nbbrd.heylogs.TimeRange;
-import nbbrd.heylogs.VersionFilter;
-import org.apache.maven.plugin.AbstractMojo;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugins.annotations.LifecyclePhase;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
 
 import java.io.File;
-import java.io.IOException;
-import java.io.Reader;
-import java.io.Writer;
-import java.nio.file.Files;
-import java.time.LocalDate;
-import java.time.format.DateTimeParseException;
 import java.util.Objects;
-import java.util.regex.Pattern;
-import java.util.regex.PatternSyntaxException;
+
+import static internal.heylogs.maven.plugin.MojoFunction.onLocalDate;
+import static internal.heylogs.maven.plugin.MojoFunction.onPattern;
 
 @Mojo(name = "extract", defaultPhase = LifecyclePhase.GENERATE_RESOURCES, threadSafe = true)
-public final class ExtractMojo extends AbstractMojo {
+public final class ExtractMojo extends HeylogsMojo {
 
     @Parameter(defaultValue = "${project.basedir}/CHANGELOG.md", property = "heylogs.input.file")
     private File inputFile;
@@ -43,11 +35,11 @@ public final class ExtractMojo extends AbstractMojo {
     @Parameter(defaultValue = "0x7fffffff", property = "heylogs.limit")
     private int limit;
 
-    @Parameter(defaultValue = "false", property = "heylogs.skip")
-    private boolean skip;
-
     @Parameter(defaultValue = "^.*-SNAPSHOT$", property = "heylogs.unreleased.pattern")
     private String unreleasedPattern;
+
+    @Parameter(defaultValue = "false", property = "heylogs.ignore.content")
+    private boolean ignoreContent;
 
     @Override
     public void execute() throws MojoExecutionException {
@@ -61,70 +53,29 @@ public final class ExtractMojo extends AbstractMojo {
             throw new MojoExecutionException("Changelog not found");
         }
 
-        getLog().info("Reading " + inputFile);
-        Document changelog = read();
-
-        VersionFilter filter = getFilter();
-
-        getLog().info("Extracting with " + filter);
-        filter.apply(changelog);
-
-        getLog().info("Writing " + outputFile);
-        write(changelog);
+        extract(loadExtractor());
     }
 
-    private VersionFilter getFilter() throws MojoExecutionException {
-        return VersionFilter
+    private Extractor loadExtractor() throws MojoExecutionException {
+        return Extractor
                 .builder()
                 .ref(Objects.toString(ref, ""))
-                .unreleasedPattern(fetchUnreleasedPattern())
-                .timeRange(TimeRange.of(fetchFrom(), fetchTo()))
+                .unreleasedPattern(onPattern("Invalid unreleased pattern").applyWithMojo(unreleasedPattern))
+                .timeRange(TimeRange.of(
+                        onLocalDate("Invalid format for 'from' parameter").applyWithMojo(from),
+                        onLocalDate("Invalid format for 'to' parameter").applyWithMojo(to))
+                )
                 .limit(limit)
+                .ignoreContent(ignoreContent)
                 .build();
     }
 
-    private Pattern fetchUnreleasedPattern() throws MojoExecutionException {
-        try {
-            return Pattern.compile(unreleasedPattern);
-        } catch (PatternSyntaxException ex) {
-            throw new MojoExecutionException("Invalid unreleased pattern", ex);
-        }
-    }
+    private void extract(Extractor extractor) throws MojoExecutionException {
+        Document changelog = readChangelog(inputFile);
 
-    private LocalDate fetchFrom() throws MojoExecutionException {
-        try {
-            return VersionFilter.parseLocalDate(from);
-        } catch (DateTimeParseException ex) {
-            throw new MojoExecutionException("Invalid format for 'from' parameter", ex);
-        }
-    }
+        getLog().info("Extracting with " + extractor);
+        extractor.extract(changelog);
 
-    private LocalDate fetchTo() throws MojoExecutionException {
-        try {
-            return VersionFilter.parseLocalDate(to);
-        } catch (DateTimeParseException ex) {
-            throw new MojoExecutionException("Invalid format for 'to' parameter", ex);
-        }
-    }
-
-    public Document read() throws MojoExecutionException {
-        Parser parser = Parser.builder().build();
-        try (Reader reader = Files.newBufferedReader(inputFile.toPath())) {
-            return parser.parseReader(reader);
-        } catch (IOException ex) {
-            throw new MojoExecutionException("Failed to read file", ex);
-        }
-    }
-
-    public void write(Document document) throws MojoExecutionException {
-        try {
-            Files.createDirectories(outputFile.getParentFile().toPath());
-            Formatter formatter = Formatter.builder().build();
-            try (Writer writer = Files.newBufferedWriter(outputFile.toPath())) {
-                formatter.render(document, writer);
-            }
-        } catch (IOException ex) {
-            throw new MojoExecutionException("Failed to write file", ex);
-        }
+        writeChangelog(changelog, outputFile);
     }
 }