+++ CHANGELOG.md
@@ -7,6 +7,22 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [0.6.0] - 2023-06-20
+
+This release improves extension points and also aligns features of Maven plugin and CLI.
+
+### Added
+
+- Add extension point for scan formatting [#119](https://github.com/nbbrd/heylogs/issues/119)
+- Add scan mojo [#120](https://github.com/nbbrd/heylogs/issues/120)
+- Add list command and mojo [#120](https://github.com/nbbrd/heylogs/issues/120)
+
+### Changed
+
+- Refactor extension points [#119](https://github.com/nbbrd/heylogs/issues/119)
+- Merge old list command into extract command [#120](https://github.com/nbbrd/heylogs/issues/120)
+- Improve output of errors in check mojo [#119](https://github.com/nbbrd/heylogs/issues/119)
+
 ## [0.5.0] - 2022-11-29
 
 ### Added
@@ -67,7 +83,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Initial release
 
-[Unreleased]: https://github.com/nbbrd/heylogs/compare/v0.5.0...HEAD
+[Unreleased]: https://github.com/nbbrd/heylogs/compare/v0.6.0...HEAD
+[0.6.0]: https://github.com/nbbrd/heylogs/compare/v0.5.0...v0.6.0
 [0.5.0]: https://github.com/nbbrd/heylogs/compare/v0.4.0...v0.5.0
 [0.4.0]: https://github.com/nbbrd/heylogs/compare/v0.3.2...v0.4.0
 [0.3.2]: https://github.com/nbbrd/heylogs/compare/v0.3.1...v0.3.2